% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{POD Cities shape file}
\format{
\subsection{\code{cities}}{

A \code{sf} data.frame with 39 rows and 4 columns.
\describe{
\item{code_muni}{Numeric code identifying each city.}
\item{name_muni}{Name of the city.}
\item{code_muni_ibge}{7-digit numeric code from IBGE.}
\item{area_ha}{Area of the polygon in hectares (ha).}
}
}
}
\source{
\url{https://www.metro.sp.gov.br/pesquisa-od/}
}
\usage{
cities
}
\description{
A spatial \code{sf} object with the city borders of all POD cities.
}
\keyword{datasets}
