% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pod_tables.R
\name{connect_pod_db}
\alias{connect_pod_db}
\title{Connect to the POD database}
\usage{
connect_pod_db()
}
\value{
A \code{SQLiteConnection} with the POD database
}
\description{
Should be used together with the \verb{pod_table_*} functions.
}
