% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_pod.R
\name{design_pod}
\alias{design_pod}
\title{Create the survey design}
\usage{
design_pod(df, id)
}
\arguments{
\item{df}{A \code{data.frame} with POD microdata.}

\item{id}{Choice of identifier. Must be one of \code{person} (person),
\code{family} (family), or \code{hh} (household).}
}
\value{
A survey object
}
\description{
Create the a survey design for the POD microdata
}
