% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts}
\alias{districts}
\title{POD Districts and Regions}
\format{
\subsection{\code{districts}}{

A \code{sf} data.frame with 134 rows and 7 columns.
\describe{
\item{code_district}{Numeric code identifying each district.}
\item{name_district}{Name of the district.}
\item{area_ha}{Area of the polygon in hectares (ha).}
\item{is_cbd}{Binary variable (dummy) indicating if district is inside the São Paulo CBD.}
\item{name_region}{Name of the region. Only applies for São Paulo}
\item{region_simplified}{Simplified name of each region.}
\item{code_region}{Numeric code identifying each region.}
}
}
}
\source{
\url{https://www.metro.sp.gov.br/pesquisa-od/}
}
\usage{
districts
}
\description{
A spatial \code{sf} object with the district borders all of POD cities.
}
\keyword{datasets}
