% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pod_tables.R
\name{import_pod_tables}
\alias{import_pod_tables}
\title{Replicate the POD tables}
\usage{
import_pod_tables(cached = TRUE, geo = FALSE, tables = "all")
}
\arguments{
\item{cached}{If \code{FALSE} creates the tables from scratch. If \code{TRUE} downloads
from Github repo.}

\item{geo}{If \code{TRUE} joins the table with the \code{zones} shape file.}

\item{tables}{String indicating which table to build. Options are \code{'demographic'},
\code{'education'}, \code{'income'}, \code{'income_group'}, \code{'cars'}, \code{'jobs'}, \code{'jobs_sector'},
or \code{'all'} (default).}
}
\value{
Either a \code{tibble} or a \code{sf} object with aggregated information for each
OD zone.
}
\description{
Replicates the main POD tables and adds some auxiliary variables
}
