% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pod_travel_table.R
\name{import_pod_travel_tables}
\alias{import_pod_travel_tables}
\title{Import POD travel tables}
\usage{
import_pod_travel_tables(cached = TRUE, geo = FALSE, tables = "all")
}
\arguments{
\item{cached}{If \code{FALSE} creates the tables from scratch. If \code{TRUE} downloads
from Github repo.}

\item{geo}{If \code{TRUE} joins the table with the \code{zones} shape file.}

\item{tables}{String indicating which table to build. Options are \code{'demographic'},
\code{'education'}, \code{'income'}, \code{'income_group'}, \code{'cars'}, \code{'jobs'}, \code{'jobs_sector'},
or \code{'all'} (default).}
}
\value{
A named \code{list}
}
\description{
Import POD travel tables
}
