% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones}
\alias{zones}
\title{POD Zones}
\format{
\subsection{\code{zones}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{code_zone}{Numeric code identifying each OD zone.}
\item{name_zone}{Name of the OD zone.}
\item{code_muni}{Numeric code identifying each city.}
\item{name_muni}{Name of the city.}
\item{code_district}{Numeric code identifying each district.}
\item{name_district}{Name of the district.}
\item{area_ha}{Area of the polygon in hectares (ha).}
\item{is_cbd}{Binary variable (dummy) indicating if district is inside the São Paulo CBD.}
\item{name_region}{Name of the region. Only applies for São Paulo}
\item{region_simplified}{Simplified name of each region.}
\item{code_region}{Numeric code identifying each region.}
}
}
}
\source{
\url{https://www.metro.sp.gov.br/pesquisa-od/}
}
\usage{
zones
}
\description{
A spatial \code{sf} data.frame with 517 rows and 11 columns.
}
\keyword{datasets}
